Add Go 1.12 and Go modules support.
This commit switches the Travis environment to use Go 1.12.x, the latest
stable release at the time of writing.

Modules support is added to the repository and dependencies are
vendored.

Lastly the `Makefile` is updated to match `zlint`'s, most notably this
means using the vendored dependencies to build and test the project.